requirements.txt
Flask>=3.1.0
Werkzeug>=3.1.0
Pillow>=11.0.0
scikit-learn>=1.7.0
numpy>=2.3.0
webcolors>=24.8.0
